Merging yasin into develop2

+++ src/main/java/com/hospitalsearch/service/impl/UserServiceImpl.java
@@ -1,12 +1,13 @@
 package com.hospitalsearch.service.impl;
 
 import com.hospitalsearch.dao.UserDAO;
-import com.hospitalsearch.dto.UserAdminDTO;
+import com.hospitalsearch.dto.UserFilterDTO;
 import com.hospitalsearch.dto.UserRegisterDTO;
 import com.hospitalsearch.entity.PatientCard;
 import com.hospitalsearch.entity.Role;
 import com.hospitalsearch.entity.User;
 import com.hospitalsearch.entity.UserDetail;
+import com.hospitalsearch.exception.ResetPasswordException;
 import com.hospitalsearch.service.PatientCardService;
 import com.hospitalsearch.service.RoleService;
 import com.hospitalsearch.service.UserService;
@@ -44,15 +45,11 @@ public class UserServiceImpl implements UserService {
     public void save(User newUser) {
         try {
             logger.info("save user: " + newUser);
-            newUser.setPassword(passwordEncoder.encode(newUser.getPassword()));
             PatientCard patientCard = patientCardService.add(new PatientCard());
             UserDetail userDetail = new UserDetail();
             userDetail.setPatientCard(patientCard);
-            userDetail.setFirstName("Anonymous");
-            userDetail.setLastName("Anonymous");
             userDetail.setGender(Gender.MAN);
             userDetail.setAddress("");
-            //todo refactor
             newUser.setUserDetails(userDetail);
             dao.save(newUser);
         } catch (Exception e) {
@@ -66,8 +63,14 @@ public class UserServiceImpl implements UserService {
         try {
             logger.info("register user: " + userRegisterDTO);
             user.setEmail(userRegisterDTO.getEmail().toLowerCase());
-            user.setPassword(userRegisterDTO.getPassword());
-            user.setUserRoles(new HashSet<>(Collections.singletonList(roleService.getByType("PATIENT"))));
+            user.setPassword(passwordEncoder.encode(userRegisterDTO.getPassword()));
+            user.setUserRoles(userRegisterDTO.getUserRoles());
+            user.setEnabled(userRegisterDTO.getEnabled());
+            if (!userRegisterDTO.getUserRoles().isEmpty()) {
+                user.setUserRoles(userRegisterDTO.getUserRoles());
+            } else {
+                user.setUserRoles(new HashSet<>(Collections.singletonList(roleService.getByType("PATIENT"))));
+            }
             save(user);
         } catch (Exception e) {
             logger.error("Error register user: " + userRegisterDTO, e);
@@ -79,23 +82,11 @@ public class UserServiceImpl implements UserService {
     public void delete(Long id) {
         User user = dao.getById(id);
         try {
+            if (isAdmin(id)) return;
             logger.info("Delete user " + user);
-            for (Role role : user.getUserRoles()) {
-                if (role.getType().equals("ADMIN")) return;
-            }
             dao.delete(user);
         } catch (Exception e) {
             logger.error("Error delete user: " + user, e);
-        }
-    }
-
-    @Override
-    public void updateUser(User user) {
-        try {
-            logger.info("Update merge " + user);
-            dao.updateUser(user);
-        } catch (Exception e) {
-            logger.error("Error merge user: " + user, e);
         }
     }
 
@@ -145,20 +136,20 @@ public class UserServiceImpl implements UserService {
         return user;
     }
 
+
     @Override
-    public User changeStatus(Long id) {
-        User user = null;
+    public void changeStatus(Long id) {
         try {
+            if (isAdmin(id)) return;
             logger.info("Change status to user with id " + id);
-            user = dao.changeStatus(id);
+            dao.changeStatus(id);
         } catch (Exception e) {
             logger.error("Error changing status user with id " + id, e);
         }
-        return user;
     }
 
     @Override
-    public boolean resetPassword(String email, String newPassword) {
+    public void resetPassword(String email, String newPassword) throws ResetPasswordException {
         User user = getByEmail(email);
         try {
             logger.info("Change password to user with email " + email);
@@ -166,30 +157,14 @@ public class UserServiceImpl implements UserService {
             dao.update(user);
         } catch (Exception e) {
             logger.error("Error changing password user with email " + email, e);
-            return false;
         }
-        return true;
     }
 
-   /* //change password
     @Override
-    public boolean changePassword(String email, String currentPassword, String newPassword) {
-        User user = getUserByEmail(email);
-
-        if (!this.passwordEncoder.matches(currentPassword, user.getPassword())) {
-            return false;
-        }
-        user.setPassword(this.passwordEncoder.encode(newPassword));
-        userDao.update(user);
-
-        return true;
-    }*/
-
-    @Override
-    public List<User> getUsers(UserAdminDTO userAdminDTO) {
+    public List<User> getUsers(UserFilterDTO userFilterDTO) {
         List<User> users = new ArrayList<>();
         try {
-            users = dao.getUsers(userAdminDTO);
+            users = dao.getUsers(userFilterDTO);
             logger.info("Get all users!");
         } catch (Exception e) {
             logger.error("Error getting all users", e);
@@ -198,28 +173,15 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public List<User> searchUser(UserAdminDTO userAdminDTO) {
+    public List<User> searchUser(UserFilterDTO userFilterDTO) {
         List<User> users = new ArrayList<>();
         try {
-            users = dao.searchUser(userAdminDTO);
+            users = dao.searchUser(userFilterDTO);
             logger.info("Search users!");
         } catch (Exception e) {
             logger.error("Error searching users", e);
         }
         return users;
-    }
-
-    @Override
-    public void registerUpdate(UserDto dto, String email) {
-    }
-
-    @Override
-    public UserDto getDtoByEmail(String email) {
-        return null;
-    }
-
-    @Override
-    public void registerUpdate(UserDetailRegisterDto dto, String email) {
     }
 
     //Illia
@@ -267,4 +229,15 @@ public class UserServiceImpl implements UserService {
         return (int) Math.ceil((double) countOfItems / itemsPerPage);
     }
     //Illia
+
+    //utilities methods
+    private boolean isAdmin(Long id) {
+        User user = dao.getById(id);
+        for (Role role : user.getUserRoles()) {
+            if (role.getType().equals("ADMIN")) {
+                return true;
+            }
+        }
+        return false;
+    }
 }
